+++ changelog/pending/20240627--cli-display--display-retain-in-all-cases-of-delete-retention.yaml
@@ -1,0 +1,4 @@
+changes:
+- type: fix
+  scope: cli/display
+  description: Display `[retain]` in all cases of delete retention
