+++ pkg/engine/lifecycletest/testdata/output/TestRetainOnDelete/0/eventstream.json
@@ -1,0 +1,7 @@
+{"sequence":0,"timestamp":0,"preludeEvent":{"config":{}}}
+{"sequence":0,"timestamp":0,"resourcePreEvent":{"metadata":{"op":"create","urn":"urn:pulumi:test::test::pulumi:providers:pkgA::default","type":"pulumi:providers:pkgA","old":null,"new":{"type":"pulumi:providers:pkgA","urn":"urn:pulumi:test::test::pulumi:providers:pkgA::default","custom":true,"id":"","parent":"","inputs":{},"outputs":{},"provider":""},"detailedDiff":null,"logical":true,"provider":""}}}
+{"sequence":0,"timestamp":0,"resOutputsEvent":{"metadata":{"op":"create","urn":"urn:pulumi:test::test::pulumi:providers:pkgA::default","type":"pulumi:providers:pkgA","old":null,"new":{"type":"pulumi:providers:pkgA","urn":"urn:pulumi:test::test::pulumi:providers:pkgA::default","custom":true,"id":"b5f6dde3-5d00-40f7-81f8-7f7496aefccc","parent":"","inputs":{},"outputs":{},"provider":""},"detailedDiff":null,"logical":true,"provider":""}}}
+{"sequence":0,"timestamp":0,"resourcePreEvent":{"metadata":{"op":"create","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","type":"pkgA:m:typA","old":null,"new":{"type":"pkgA:m:typA","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","custom":true,"id":"","parent":"","retainOnDelete":true,"inputs":{"foo":"bar"},"outputs":{},"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"},"detailedDiff":null,"logical":true,"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"}}}
+{"sequence":0,"timestamp":0,"resOutputsEvent":{"metadata":{"op":"create","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","type":"pkgA:m:typA","old":null,"new":{"type":"pkgA:m:typA","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","custom":true,"id":"created-id-0","parent":"","retainOnDelete":true,"inputs":{"foo":"bar"},"outputs":{"foo":"bar"},"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"},"detailedDiff":null,"logical":true,"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"}}}
+{"sequence":0,"timestamp":0,"summaryEvent":{"maybeCorrupt":false,"durationSeconds":1,"resourceChanges":{"create":1},"PolicyPacks":{}}}
+{"sequence":0,"timestamp":0,"cancelEvent":{}}
