+++ pkg/engine/lifecycletest/testdata/output/TestRetainOnDelete/1/eventstream.json
@@ -1,0 +1,11 @@
+{"sequence":0,"timestamp":0,"preludeEvent":{"config":{}}}
+{"sequence":0,"timestamp":0,"resourcePreEvent":{"metadata":{"op":"same","urn":"urn:pulumi:test::test::pulumi:providers:pkgA::default","type":"pulumi:providers:pkgA","old":{"type":"pulumi:providers:pkgA","urn":"urn:pulumi:test::test::pulumi:providers:pkgA::default","custom":true,"id":"b5f6dde3-5d00-40f7-81f8-7f7496aefccc","parent":"","inputs":{},"outputs":{},"provider":""},"new":{"type":"pulumi:providers:pkgA","urn":"urn:pulumi:test::test::pulumi:providers:pkgA::default","custom":true,"id":"","parent":"","inputs":{},"outputs":{},"provider":""},"detailedDiff":null,"logical":true,"provider":""}}}
+{"sequence":0,"timestamp":0,"resOutputsEvent":{"metadata":{"op":"same","urn":"urn:pulumi:test::test::pulumi:providers:pkgA::default","type":"pulumi:providers:pkgA","old":{"type":"pulumi:providers:pkgA","urn":"urn:pulumi:test::test::pulumi:providers:pkgA::default","custom":true,"id":"b5f6dde3-5d00-40f7-81f8-7f7496aefccc","parent":"","inputs":{},"outputs":{},"provider":""},"new":{"type":"pulumi:providers:pkgA","urn":"urn:pulumi:test::test::pulumi:providers:pkgA::default","custom":true,"id":"b5f6dde3-5d00-40f7-81f8-7f7496aefccc","parent":"","inputs":{},"outputs":{},"provider":""},"detailedDiff":null,"logical":true,"provider":""}}}
+{"sequence":0,"timestamp":0,"resourcePreEvent":{"metadata":{"op":"create-replacement","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","type":"pkgA:m:typA","old":{"type":"pkgA:m:typA","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","custom":true,"id":"created-id-0","parent":"","retainOnDelete":true,"inputs":{"foo":"bar"},"outputs":{"foo":"bar"},"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"},"new":{"type":"pkgA:m:typA","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","custom":true,"id":"","parent":"","retainOnDelete":true,"inputs":{"foo":"baz"},"outputs":{},"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"},"keys":["foo"],"detailedDiff":null,"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"}}}
+{"sequence":0,"timestamp":0,"resOutputsEvent":{"metadata":{"op":"create-replacement","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","type":"pkgA:m:typA","old":{"type":"pkgA:m:typA","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","custom":true,"delete":true,"id":"created-id-0","parent":"","retainOnDelete":true,"inputs":{"foo":"bar"},"outputs":{"foo":"bar"},"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"},"new":{"type":"pkgA:m:typA","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","custom":true,"id":"created-id-1","parent":"","retainOnDelete":true,"inputs":{"foo":"baz"},"outputs":{"foo":"baz"},"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"},"keys":["foo"],"detailedDiff":null,"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"}}}
+{"sequence":0,"timestamp":0,"resourcePreEvent":{"metadata":{"op":"replace","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","type":"pkgA:m:typA","old":{"type":"pkgA:m:typA","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","custom":true,"delete":true,"id":"created-id-0","parent":"","retainOnDelete":true,"inputs":{"foo":"bar"},"outputs":{"foo":"bar"},"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"},"new":{"type":"pkgA:m:typA","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","custom":true,"id":"created-id-1","parent":"","retainOnDelete":true,"inputs":{"foo":"baz"},"outputs":{"foo":"baz"},"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"},"keys":["foo"],"detailedDiff":null,"logical":true,"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"}}}
+{"sequence":0,"timestamp":0,"resOutputsEvent":{"metadata":{"op":"replace","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","type":"pkgA:m:typA","old":{"type":"pkgA:m:typA","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","custom":true,"delete":true,"id":"created-id-0","parent":"","retainOnDelete":true,"inputs":{"foo":"bar"},"outputs":{"foo":"bar"},"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"},"new":{"type":"pkgA:m:typA","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","custom":true,"id":"created-id-1","parent":"","retainOnDelete":true,"inputs":{"foo":"baz"},"outputs":{"foo":"baz"},"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"},"keys":["foo"],"detailedDiff":null,"logical":true,"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"}}}
+{"sequence":0,"timestamp":0,"resourcePreEvent":{"metadata":{"op":"delete-replaced","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","type":"pkgA:m:typA","old":{"type":"pkgA:m:typA","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","custom":true,"delete":true,"id":"created-id-0","parent":"","retainOnDelete":true,"inputs":{"foo":"bar"},"outputs":{"foo":"bar"},"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"},"new":null,"detailedDiff":null,"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"}}}
+{"sequence":0,"timestamp":0,"resOutputsEvent":{"metadata":{"op":"delete-replaced","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","type":"pkgA:m:typA","old":{"type":"pkgA:m:typA","urn":"urn:pulumi:test::test::pkgA:m:typA::resA","custom":true,"delete":true,"id":"created-id-0","parent":"","retainOnDelete":true,"inputs":{"foo":"bar"},"outputs":{"foo":"bar"},"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"},"new":null,"detailedDiff":null,"provider":"urn:pulumi:test::test::pulumi:providers:pkgA::default::b5f6dde3-5d00-40f7-81f8-7f7496aefccc"}}}
+{"sequence":0,"timestamp":0,"summaryEvent":{"maybeCorrupt":false,"durationSeconds":1,"resourceChanges":{"replace":1},"PolicyPacks":{}}}
+{"sequence":0,"timestamp":0,"cancelEvent":{}}
